Add the basics of keyboard navigation and widget focus...
There is a lot of unexplored territory here and refinement to be had.
Furthermore a fair amount of widgets will need to implement
custom focus functionality so that the focus moves as the user
would expect.

The next big thing is to decide what kind of focus we will support.
Is the mouse and keyboard focus going to be different?
How will they interact?
Will we have soft focus? etc.