Add support for special keyboard shortcuts to bypass tab capturing...
This is needed since you might want to add a binding for tab like
in the LineEdit implementation.
Because the keyboard navigation shortcuts get evaluated at the end,
specifically to allow overrides, we need something extra that gets
evaluated before anything else and uses less common keys to reduce
the likeliness of a clash with keybinds people might want to set.