Finish 90% implementation of focus and keyboard navigation...
We now distinguish between soft and hard focus.
Soft focus is when you traverse the UI using the keyboard.
Widgets that are soft focused do not respond to keyboard events unless
you press the space key. Which will "activate" the Widget changing it
to hard focus.
Hard focus is equivalent to the old focus and comes either from a
mouse down event or from an activate call (whether programmatic or space).
A Widget that is hard focused will be evaluated against keybindings etc.

Another upside from implementing activate() is that we can
programmatically select a Widget (press a button for example) and have
something happen (like activating a button to change tabs which will load different data as in mhwidb).

Lastly the reason its 90% is that we havent implemented "nice" propagation
for focus events for all the provided Widgets.
Once thats done it should be 100%.