Added keyboard navigation for some widgets (including notebook!)...
One thing I'm unsure of is isFocusable for layouts to get focus events
in the handler. It works well for the most part, and we can override
for some widgets where it causes some small undesirable behavior.
But the biggest (still small) issue is that it makes it so you can
focus a layout using the mouse, which is not great. Most likely
I will need to add something to distinguish layouts that don't actually
focus but still need the event.