Initial support for multiple window...
This is not quite done yet, there are a few kinks to work out but it
is already usable.
The biggest TODO item regarding multiple windows is to investigate and
potentially use shared OpenGL contexts in order to be able to use
the same fonts and images across all windows.
Otherwise there are various random things left to do like
only redrawing the windows which need an update rather than all of them.