Fix built in icons for multiple windows...
I explored shared contexts a bit so that we could use a single context.
That particular approach didnt quite work. There was a problem with
the window being currently drawn to affecting the drawable region of the
rest. In the end I ended up going with making each context sharable.
Each window has its own context but any textures are shareable between them.