Update focus handling...
In cases where we pass the event to another widget we should always
pass the original event data. In this case it doesnt matter but better
to keep it consistent.

When a widget after the notebook wants to go in reverse we should
go to the last widget of the current tab.