Added a widget for editing styles with an option to use it as a separate window...
It is mostly finished. However there is some work surrounding it that
needs to be done. For example we need to make sure that forEachWidget
actually does go over every widget so we cant force update the style
information. Some widgets dont fully respond to style but that will
be cleaned up in due time.