Generalize DropDownList into List and restrict it to CellRenderers...
I decided to use strictly CellRenderers rather than Drawables
because Drawable would mean you can use Widgets and I feel like that
doesnt make sense within this Widget.
All Widgets should be interactive to some extent. They should be able
to respond to events and receive keybindings etc.
This would not have been honored within a list.
Furthermore using Widgets vastly over complicates the drawing of the
current item in the DropDown and it makes it harder to show
selection and hover effects.

Because the DropDownList was already its own thing I made the few
changes necessary to make it fully standalone and just made
DropDown interact with it.