Remove the need for some widgets to call sizeHint on themselves before drawing...
This was necessary because some widgets, especially layouts, require
the information computed by sizehint in order to function properly.
This is because they need size information about all their children
for example. Same case with treeview.
Instead of individual widgets being responsible for it
I added a call to sizehint on the main widget within a window draw call.
There should be no performance change since sizehint was already being
called by the widgets themselves.
This should just make it more convenient for other widgets that
require the same kind of pattern.