Stop recreating scrollbars in favour of showing them as needed...
Previously we had to delete scrollbars when they werent needed
and create them again when they were.
I switched it to use the widget's hide and show methods instead.
This is makes more sense as it the functionality is already there
and it means we are not constantly allocating scrollbars for no reason.
The reason why we didnt do this earlier is because scrollbars can cause
a call to sizehint before we have an application, since an application is
a window and gets constructed first and the default main widget for a
window is a scrolledbox which has scrollbars.
This is why we construct the scrollbars in the scrollable sizehint
since at that point we can guarantee that we can access an instance
of a drawing context.