Add basic unicode support...
This has been a long time coming :)
Its mostly ok for now. Text handling in LineEdit will need to be updated
accordingly since it still operates on bytes rather than codepoints.
In the future we should add support for proper glyph composition
and text handling for things that use more than one codepoint
like flag emoji or arabic etc but for now this is sufficient.

Revamped font to pre-load ascii glyphs and load in others as needed.
The initial load only uses one loop so font loading should have become faster.
If the font atlas texture is not big enough a new one is created.
Atm the growth process is suboptimal see comment when we grow.
We could eventually run into an issue where the opengl texture is too
big so thats something to consider in the future.

The relevant code in DrawingContext and Renderer has been updated to
operate on utf8 sequences rather than individual bytes (there are some
files that still need to be updated for that).

Updated code that checked for the height of the text to use the new API.