Decoupled Application from Window in favour of composition...
I've always wanted to make that change but it is less convenient.
Though for the most part is just means that instead of app you say
app->mainWindow().

Made the necessary changes to prevent segfaults since the order of
constructors and destructors has changed. We also moved the SDL
initialisation into Application as part of the constructor which
required some things like cursors to be moved to the heap since we cant
initialise it before SDL.

This change has been mostly spurned by segfaults when freeing fonts.
The default font would never cause this because it was never freed.
It didnt really have to be because at that point we were shutting down
the application anyway. But in essence because of the dtor ordering
we were freeing the freetype library before the individual font faces.
Now the default font is freed and things are freed in the correct order.

The only problem now is that SDL timers will only get cleanup when
SDL is being free... which is after the Windows have been destroyed.
So the plan for that is to introduce an abstraction for the timers
since at the moment we use SDL directly and just kill the timers before
the windows. This should also allows us to use lambdas for timers which
would be quite convenient.