Add string concatenation by utilizing insert...
Do note however that the string copying and allocations are pretty bad.
Part of that is how we join the strings and part is how we grow the
capacity. But i think this is probably fine, and if anything we would
be better of just writing a dedicated string builder or something.