Added TextEdit which will eventually replace LineEdit...
For now this is just a direct copy but swapping the string type.
TextEdit will be a multiline widget that optionally can function as
a form entry which doesnt capture tab and or a single line text widget.

Since we will not use this for a code editor and I dont have plans to
write a sophisticated text buffer... for now this will just use an
array of lines implementation.

At the same time the goals are to:
fix outstanding history bugs
improve scrolling
and mouse selection especially in more complex situations like
the text widget being part of treeview cell.
Finally to make sure we handle all text as codepoints (grapheme
clusters will have to wait) at least when appropriate.