Disable some functionality for now while we implement the basics...
Also change the way that sizehint works since we want the component
to support both single line and multi line modes.
Same with text and setText.

Next up we will update the way in which we draw text and to make sure
it works well with scrolling.
For single line mode we might omit showing the scrollbars tbd.