Fix bug where we would slice into a utf8 codepoint sequence...
This was caused by incorrect usage of the data field on the iter.
For the most part it is not meant to be used directly anyway.
The lesson for the future is to just have a measureText version
that takes a codepoint directly rather than have to make a slice.
In that case we have never run into this issue.