Basic multiline selection working and rendering...
For the moment we only support the use case where the beginning is
always before the end. We might be able to just get away with swapping
as thats what we did in the past but it is important to preserve the
original selection for undo redo events at the very least.