Reimplement move cursor left and right for TextEdit...
This is a much nicer implementation that the previous one...
having a second stab at something is always much better.
On top of that we now move to the next line if we reach the end of the
current line (one below or one above depending on which way we are moving).

One thing to mindful of for the near future is that we still need
a nice way to deal with newlines.
Here we might want to just skip over them, or perhaps not event have
them as part of the buffer and only insert them when the user
asks for the whole text buffer. Will need to decide on that.
And on a related not we might want to extend the selection slightly
if it goes over into the next line to show that we are deleting the whole
line (once that is supported anyway).