Basic implementation for (only Y atm) scrolling with the mouse...
This was a surprisingly difficult solution to come to (maybe i was just tired).
It works fine atm, we just need to add the same for X axis.
However i think we will try to adjust it such that we have a default low
speed for the scrolling that gets some acceleration depending on how much
the user has moved the mouse, as this constant speed doesnt feel that good
on larger documents.