Fix text rendering outside of its rect...
This was especially noticeable on big font sizes with unicode text.
Essentially the problem was some unicode glyph have bearing that is
different enough that when we used 'H' as the base bearing it would make
us start drawing higher than we initially wanted.
The fix is simple. Just start drawing account for the max bearing rather
than using 'H' as a baseline (which we were only using because thats
what learnopengl materials used).