Fix single deletion when it would cause a merge with the next line...
No need to skip over the newline anymore since they are not in the buffer.
Update virtual_size height since we werent doing that, although failing
to do that is not really disastrous since it only affects the viewport.
Correctly check if we there is a next line using the current last line + 1.