Correct condition which determines whether we draw placeholder text or not...
The previous condition would sometimes cause the widget to draw placeholder text
when it wasnt supposed to. This happened because we checked the first line for whether
it was empty but we did that even when there was more than one line of text :(
The reason this used to be fine before is that we actually stored the newline
characters in the lines themselves so in effect the lines were never empty.