Implement deletions for selections spanning multiple lines...
At this point the only thing to implement with deletions is backspace
but that would only affect deletions with no selection.
So rather than delete on codepoint to the right we delete one to the left
and if we hit the beginning of a line that we merge the current one
with the one above.