Slight rework of insert...
It is a bit simpler to reason about since singline mode is a separate branch.
This fixes any issues that I have noticed especially related to pasting in multiline text
I decided to update virtual width once at the end since it should be reasonably fast
and until it proves to be a bottle neck it is too much of a pain to deal with incrementally
especially since most of my testing now happens on the sqlite3.c file which is about 250k LOC
and it hasnt proven a problem yet.