Implement recording history and undo / redo functionality...
Finally the TextEdit is on par with the LineEdit in terms of features
so we can get rid of it.
There are a couple things to cleanup like the scrollbars and the viewport.
Essentially we have some padding that moves the scrollbars inward
leaving this weird gap between the scrollbars and the actual edges
of the viewport.
Furthermore I want to implement styling per range of characters
this way we could implement syntax highlighting for example but I need
to spend some time thinking about that first.
We are not interested in word wrap at the moment.
Not directly TextEdit but related, we need to figure out how to
communicate to the app when font size changes and other styling changes
that affect layout and size calculations so that we can update appropriately.
Also there is a slight bug where if we start scrolling with mouse movement
and move the mouse into the bottom right corner when the scroll for both axis
should be active instead they both stop (This happens because mouse event goes to nullptr).

In the next commit we should remove LineEdit from the code base
and replaces all instances with TextEdit in Singleline mode.