Remove LineEdit in favour of TextEdit...
In the future we will need to remove std::string in favour of our own String implementation.
TextEdit needs one more small (i hope lol) change before we merge into master.
Namely it encodes the inner_rect.x position into x_begin and x_end
which means that if the layout ever changes the cursor position is fucked.
In order to fix that we should only include inner_rect.x in drawing and
mouse event handling, everywhere else the value should not be included.

Lastly, now that I remove line spacing it is very obvious that when a font
grows because of new (taller) added glyphs we need to recompute size
so we probably want to revisit that overridable layout behaviour
so that we can correctly set text_change variable.