Make the layout respond to when font size changes...
This is a temporary solution that will eventually be replaced with
something like handleStyleChange where widgets could respond to
various styling changes as it makes sense for them and re-layout or not
as appropriate.

For right now we are going with this. Unfortunately when you specify
a font to render at 14px size it doesnt meant thats how tall it will
be :^)
so in order to update our widgets I'm going with this solution.
The text changed variable can at least for now live in widget since
each widget has a font associated with it anyway but i think
ideally font goes to style and we should be able to respond to text
size changes with what i mentioned in the first paragraph.