Massive changes to style and how widgets are notified of styling changes...
Previously we had a problem of widgets not responding dynamically to styling changes.
This was because of a couple reasons.
1. We did not have setters for style values which meant that you had to
manually keep track of what changed and whether that required a widget to update or not.
2. Even when we knew the widget had to update we only had the basic m_size_changed
variable to indicate to widgets that they need to redo their size calculations.
This doesnt work for some of the more complex widgets like TextEdit or TreeView.
For them the style changes can affect their virtual size (but doesnt have to!)
which needs to be recalculate and is usually a lot more expensive.

What still remains is going over all widgets and making sure they only
update their sizes as required and not on any basic change.
The next commit will include an implementation of that for TextEdit.
But some of the other ones that need to be looked at include:
TreeView
ScrolledBox
And for quite a lot of widget we could restructure sizeHint such that
when padding, margins or border change we dont recalculate the whole thing
but just the difference, since these three should be fairly easy to isolate.