Update CellRenderer for custom String and style changes...
Removed any redundant style related members and replaced them to just
use the style member of Drawable.
Which means its much easier to support padding now (any padding!)
and its response to style changes!

What remains is making accessor methods for the various CellRenderers
such that the size gets update when setting new text or image etc.

And finally it might be a good idea for style to have members for
alignment. Especially for text but to a lesser extent images as well.
Something to think about.