Correctly update all tabs when styling changes...
I also went ahead and upgraded the implementation of the active tab
border.
Previously all of this was relying on hardcoded values that wouldnt change.
Now we use either a value based on the default style or a minimum width
whichever is more applicable.
Furthermore, the border top never changes we just dont draw the border
for tabs that arent active.