Some improvements to text rendering by not drawing text glyps outside the current clip (worst case window size)
Removed window* from renderer since we could use the clipping rect for the same purpose
since it is at least the window size and possibly smaller.
Removed old functions signatures that were left in as an oversight since
the utf8 functions are part of string now and have already been removed from
the renderer in the past (but only implementation file).
Added a case to stop drawing early for the drawingcontexts filltextmultilinealigned
as it was doing the alignement itself and previously did not benefit from
renderers optimization to stop drawing based on y position.