Change fonts to use a 2D texture array instead of a normal 2D texture...
This makes it much nicer to render text when we have: a lot of different
glyphs and or at a really big font size.
The alternative would have been to string together multiple textures per
font and bind them according to the glyph being drawn.
This way a font should be able to be drawn with only one texture binding
operation since the texture space that we have available per font is now
w * h * d which on the M1 using opengl 3.3 is 16384 * 16384 * 2048.
And before this change I was running into issue where in the utf8 test
file with a decently big font we would hit the texture size limit
(although thats partially because we only really utilize the texture width
since that was easier in the beginning) which resulted in no text being drawn
and an opengl error.