Make changes so that changing scale doesnt make the UI jumpy...
When changing the font size often times with non ascii text the pixel size
of the font is not the height of the tallest character but until we load
such a character we wont know.
This means that the UI can quickly update twice but at different font sizes
causing this jump of various UI elements as their layout and size gets affected.
To prevent it we temporarily freeze the swap_buffer operation.
This means that we still do all the things that are involved with drawing
but only present changes once the freeze is lifted.

Ideally there will be a better API for this that doesnt rely on hardcoding and
when frozen we should do the smallest amount of work possible to produce
accurate size and layout results, so ideally we should not be submitting any quads
to the renderer.