Changed pulse() method...
Made it static and belong to the Application since it didnt need
anything internal and it is a global function to force the event loop
to run.
Also made it accept a LayoutEvent argument for greater flexibility
when running the event loop, defaults to LAYOUT_NONE so the behaviour
is exactly the same as before.