Make minimap into a custom slider widget to control scrolling...
This works and we handle all except one case where we need to sync
the scroll to the vertical scrollbar but for that one we probably want
to override the default scrollbars which is something we want to do anyway.

We need to have another look at drawing selection since we probably want
the minimap to handle that.
Furthermore we need to customize the slider look of the minimap for better
legibility.