Fix line drawing for more complex hierarchies...
When trying to list directories using the new fs api and outputting
that information to a treeview some lines would not be drawn when
the parent and its children were not currently within the viewport.
Furthermore in some cases when only leaf nodes were visible the lines
wouldnt be drawn either since that branch never had a check.

Now we recursively draw the lines down from each parent for all visible
nodes (subject to having a parent or higher sibling and those being
outside the viewport).
The above should never have been mutually exclusive with having a
lower sibling.