Fix ArrayList implementation for non trivially constructible types...
The implementation itself was mostly ok, issues occured when the array
was assigned or returned because of the default copy and or move ctors.

So ive implemented copy and move ctors and assignment as well as switched
the implementation to use new and delete instead of malloc and free
as this way the hidden object construction is easier to reason about
and this is seemingly necessary?? on windows.