Add a user callback that gets invoked on the main gui thread...
This is practically necessary for multithreaded applications such as
a music player. For one OpenGL requires using a thread that has set
the current context which the user should never have to do.
But even in the simplest use case of updating a progress bar using this
callback makes it much less error prone (though destroying a widget
before the callback is called is still possible).