Introduce initial API for app (ie non os) drag and drop events...
This implementation requires the widget implementer to handle the event
and draw the appropriate indication of how and where the drop occurs.
It is also the user (or the widget implementers) responsibility to call
startDrag to initiate the event. Ideally this will happen in an
onMouseDown callback or similar so default implementations can be turned off.