Fix returning the wrong value...
We always need to return the result of propagateMouseEvent since
that is the last widget that dealt with event (occasionally null... tbf)
and in this case the wrong return was preventing use from setting a context menu
active on widgets within a scrolledbox but only if there was another layout
in between so:
ScrolledBox -> Box -> Button doesnt work but
ScrolledBox -> Button does work.
Now it behaves as it should.