Expand tree view drag and drop...
Some changes might still happen but essentially:
Each widget supporting drag and drop could provide a DropAction struct.
This would provide extra widget specific information about the drop
event when notifying onDragDropped listeners.
Atm the TreeView just simply holds the struct itself but ideally
we will pass it into the notify event.
However this might need to be a void* which isnt ideal.
I'm not seeing a way on how to do it with templates and an interface
ptr wouldnt help since its widget specific.

Furthermore the TreeView now shows where the drop will happen.
You can drop things above, below or as a child of the node, or into
the root of the model.
However I think we will need to introduce something that restricts
the drop drawing because not every TreeView instance will want to support
every kind of drop and in the event listener its easy to ignore specific
kinds of DropAction types but the drawing would be misleading to the user.