Expand drop implementation for more nuanced drawing and hit testing...
By setting the drop_allow variable you can now control which drop types
get drawn and how.

For example by only specifying 0b1100 or 0b1101 (since root doesnt matter)
the hit testing for the above and below gets changed to half the height
of a node (because child is not a valid drop). The drawing stays the
same in this case because I found only drawing at the edges of the node
made the two kinds of drops more distinguishable from each other.

By only allowing the child and or root drop, the child drop subsequently
takes up the whole node in terms of hit testing and drawing.

Otherwise the drops are:
top 1/4 == above
middle 2/4 == child
bottom 1/4 == below
even if above or below (but not both) are disabled.